pyproject.toml
[tool.poetry]
name = "carrotsh"
version = "1.4.2"
description = "A lightweight and secure remote access server that allows clients to connect through a browser"
authors = ["AnnikaV9 <carrot.moncher@gmail.com>"]
license = "Unlicense"

[tool.poetry.dependencies]
python = "^3.7.0, ^3.8.0, ^3.9.0, ^3.10.0"
pyyaml = "^6.0.0"
pyotp = "^2.6.0"
cryptography = "^37.0.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
